FineGym zero-shot Table-3: Ti-FAD re-implementation and 4-level hierarchy champion
Shared ActionFormerWithCLIP backbone + frozen CLIP text pathway
(libs/modeling/text_pathway.py) with three configs: name-embedding baseline,
Ti-FAD re-implementation (per-level text-video cross-attention + fg head),
and the 4-level champion (11-sentence hierarchical prompt ensemble,
action-phase attention, apparatus- and element-set-level auxiliary losses
with 3-way learned uncertainty weighting, duration prior).

Includes the prompt ensemble, verbalizer, attribute table with the 20
held-out elements, duration prior, DeepSpeed bf16 config, the finegym_slide
raw-frame dataset updates (7-digit JPG names, nearest-frame snapping,
metadata cache), the transformers-4.34 non-reentrant gradient-checkpointing
fallback, a CPU verification script, and README_FINEGYM.md.

+++ libs/modeling/blocks.py
@@ -47,9 +47,11 @@ class MaskedConv1D(nn.Module):
         # compute the mask
         if self.stride > 1:
             # downsample the mask using nearest neighbor
+            # (interpolate lacks a BFloat16 CUDA kernel in torch < 2.1; the mask is
+            #  0/1 so float32 nearest is identical — cast back to keep x.dtype)
             out_mask = F.interpolate(
-                mask.to(x.dtype), size=out_conv.size(-1), mode='nearest'
-            )
+                mask.to(torch.float32), size=out_conv.size(-1), mode='nearest'
+            ).to(x.dtype)
         else:
             # masking out the features
             out_mask = mask.to(x.dtype)
@@ -426,7 +428,11 @@ class LocalMaskedMHCA(nn.Module):
 
     @staticmethod
     def _mask_invalid_locations(input_tensor, affected_seq_len):
-        beginning_mask_2d = input_tensor.new_ones(affected_seq_len, affected_seq_len + 1).tril().flip(dims=[0])
+        # tril/triu are not implemented for BFloat16 on CUDA (torch < 2.1); build the
+        # mask pattern in float32 (device preserved) — it is only used via `== 1`.
+        beginning_mask_2d = input_tensor.new_ones(
+            affected_seq_len, affected_seq_len + 1, dtype=torch.float32
+        ).tril().flip(dims=[0])
         beginning_mask = beginning_mask_2d[None, :, None, :]
         ending_mask = beginning_mask.flip(dims=(1, 3))
         beginning_input = input_tensor[:, :affected_seq_len, :, : affected_seq_len + 1]
